emacs: fix `notmuch-*-filter-by-tag' query strings
`notmuch-tree-filter-by-tag' and `notmuch-search-filter-by-tag' are
described, in their docstrings, as matching all the messages in the
current view that are ALSO tagged by a given tag.  However, since the
current view's search query was not logically grouped together (with
parentheses), the actual behavior was different for sufficiently
complex search queries.

Fix this by grouping the current view's search query in parentheses
first.